fix few issues
