+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/display/MiniAppDisplayViewModel.kt
@@ -1,23 +1,12 @@
 package com.rakuten.tech.mobile.testapp.ui.display
 
 import android.content.Context
-import android.content.DialogInterface
-import android.view.LayoutInflater
 import android.view.View
 import androidx.lifecycle.*
-import androidx.recyclerview.widget.DividerItemDecoration
-import androidx.recyclerview.widget.LinearLayoutManager
 import com.rakuten.tech.mobile.miniapp.MiniApp
 import com.rakuten.tech.mobile.miniapp.MiniAppDisplay
-import com.rakuten.tech.mobile.miniapp.MiniAppInfo
 import com.rakuten.tech.mobile.miniapp.MiniAppSdkException
 import com.rakuten.tech.mobile.miniapp.js.MiniAppMessageBridge
-import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermission
-import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermissionResult
-import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermissionType
-import com.rakuten.tech.mobile.miniapp.testapp.databinding.ListCustomPermissionBinding
-import com.rakuten.tech.mobile.testapp.ui.permission.CustomPermissionAdapter
-import com.rakuten.tech.mobile.testapp.ui.permission.CustomPermissionDialog
 import com.rakuten.tech.mobile.testapp.ui.settings.AppSettings
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -61,88 +50,6 @@ class MiniAppDisplayViewModel constructor(
 
     fun setHostLifeCycle(lifecycle: Lifecycle) {
         this.hostLifeCycle = lifecycle
-    }
-
-    fun promptCustomPermission(
-        context: Context,
-        miniAppInfo: MiniAppInfo,
-        permissions: List<Pair<MiniAppCustomPermissionType, String>>,
-        callback: (grantResult: String) -> Unit
-    ) {
-        val layoutInflater = LayoutInflater.from(context)
-        val permissionLayout = ListCustomPermissionBinding.inflate(layoutInflater, null, false)
-        permissionLayout.listCustomPermission.layoutManager = LinearLayoutManager(context)
-        val adapter = CustomPermissionAdapter()
-
-        permissionLayout.listCustomPermission.addItemDecoration(
-            DividerItemDecoration(
-                context,
-                DividerItemDecoration.VERTICAL
-            )
-        )
-
-        // prepare adapter for showing items
-        val namesForAdapter: ArrayList<MiniAppCustomPermissionType> = arrayListOf()
-        val resultsForAdapter: ArrayList<MiniAppCustomPermissionResult> = arrayListOf()
-        val descriptionForAdapter: ArrayList<String> = arrayListOf()
-
-        permissions.forEach {
-            namesForAdapter.add(it.first)
-            descriptionForAdapter.add(it.second)
-        }
-
-        val cachedList = miniapp.getCustomPermissions(miniAppInfo.id)?.pairValues
-
-        val filteredPair =
-            mutableListOf<Pair<MiniAppCustomPermissionType, MiniAppCustomPermissionResult>>()
-
-        permissions.forEach { (first) ->
-            if (cachedList != null) {
-                if (first == MiniAppCustomPermissionType.UNKNOWN)
-                    filteredPair.add(
-                        Pair(
-                            first,
-                            MiniAppCustomPermissionResult.PERMISSION_NOT_AVAILABLE
-                        )
-                    )
-                filteredPair.addAll(cachedList.filter {
-                    first.type == it.first.type
-                })
-            } else {
-                if (first == MiniAppCustomPermissionType.UNKNOWN)
-                    filteredPair.add(
-                        Pair(
-                            first,
-                            MiniAppCustomPermissionResult.PERMISSION_NOT_AVAILABLE
-                        )
-                    )
-                else
-                    filteredPair.add(Pair(first, MiniAppCustomPermissionResult.DENIED))
-            }
-        }
-
-        filteredPair.forEach {
-            resultsForAdapter.add(it.second)
-        }
-        adapter.addPermissionList(namesForAdapter, resultsForAdapter, descriptionForAdapter)
-        permissionLayout.listCustomPermission.adapter = adapter
-
-        val objToStore = MiniAppCustomPermission(miniAppInfo.id, adapter.permissionPairs)
-
-        val listener = DialogInterface.OnClickListener { _, _ ->
-            val response = miniapp.setCustomPermissions(
-                objToStore
-            )
-            callback.invoke(response)
-        }
-
-        val permissionDialogBuilder =
-            CustomPermissionDialog.Builder().build(context, miniAppInfo.displayName).apply {
-                setView(permissionLayout.root)
-                setListener(listener)
-            }
-
-        permissionDialogBuilder.show()
     }
 
     fun canGoBackwards(): Boolean =
